fix(fraud): correct refs.bin layout and float64 query vectorization
The prepare tool was writing one label byte after each vector row while
LoadStore expects all vectors followed by all labels, which misaligned
reads from row 1 onward and broke kNN detection (~54% failure rate).

- Write labels in a single block at end of refs.bin (RNF1 float32)
- Vectorize queries with float64 + Round4 to match data-generator C
- MCCTable uses float64 risks; SearchForestK and brute helpers for tests
- Add dataset checksum test and optional test-data sample validation

+++ internal/fraud/partition.go
@@ -24,7 +24,7 @@ func partitionFromRequest(r *Request) int {
 	return partitionTag(unknown, r.LastTransaction != nil)
 }
 
-func partitionFromVector(points []float32, dim, row int) int {
+func partitionFromVector(points []float64, dim, row int) int {
 	base := row * dim
 	unknown := points[base+11] >= 0.5
 	hasLast := points[base+5] != -1
